Moves b58 encoding to a separate file, and stubs out a few unit tests
